doctor: put the FIFO test behind a build constraint, not a runtime skip
test-windows failed to BUILD on a8f907533:

    cmd\entire\cli\doctor_test.go:1779:20: undefined: syscall.Mkfifo

syscall.Mkfifo does not exist on Windows, so guarding it with
`if runtime.GOOS == windowsGOOS { t.Skip(...) }` was the wrong tool: a runtime
skip still has to compile. Moved to doctor_fifo_unix_test.go behind `//go:build
unix`, which is the guard that actually applies, and the now-redundant runtime
skip is gone with it.

`GOOS=windows go vet ./cmd/entire/cli/` reproduces the failure in a second and
would have caught it before the push; `mise run test:ci` cannot, because it
compiles for the host only. Worth reaching for on any change that touches
syscall, build tags or platform-specific paths — the whole tree now type-checks
clean for windows, linux and darwin.

Entire-Checkpoint: 01M1X8T65FGPEZTK7W58F1PJVR
